change: engine refactor to separate interface from internal classes
