+++ server/server-engine/src/main/java/com/thegame/server/engine/intern/services/impl/ChatServiceImpl.java
@@ -1,9 +1,9 @@
-package com.thegame.server.engine.services.impl;
+package com.thegame.server.engine.intern.services.impl;
 
 import com.thegame.server.engine.messages.ChatMessageBean;
-import com.thegame.server.engine.services.BusinessServiceFactory;
-import com.thegame.server.engine.services.ChatService;
-import com.thegame.server.engine.services.PlayerService;
+import com.thegame.server.engine.intern.BusinessServiceFactory;
+import com.thegame.server.engine.intern.services.ChatService;
+import com.thegame.server.engine.intern.services.PlayerService;
 import java.util.stream.Stream;
 
 /**
