+++ server/server-engine/src/main/java/com/thegame/server/engine/intern/tasks/BaseMessageTask.java
@@ -1,4 +1,4 @@
-package com.thegame.server.engine.tasks;
+package com.thegame.server.engine.intern.tasks;
 
 import com.thegame.server.common.logging.LogStream;
 import lombok.Getter;
