+++ server/server-engine/src/main/java/com/thegame/server/engine/intern/tasks/RegisterPlayerTask.java
@@ -1,9 +1,9 @@
-package com.thegame.server.engine.tasks;
+package com.thegame.server.engine.intern.tasks;
 
-import com.thegame.server.engine.data.PlayerData;
+import com.thegame.server.engine.intern.data.PlayerData;
 import com.thegame.server.engine.messages.RegisterPlayerMessageBean;
-import com.thegame.server.engine.services.BusinessServiceFactory;
-import com.thegame.server.engine.services.PlayerService;
+import com.thegame.server.engine.intern.BusinessServiceFactory;
+import com.thegame.server.engine.intern.services.PlayerService;
 
 /**
  * @author afarre
